+++ packages/app/src/actions/__tests__/index.spec.js
@@ -32,8 +32,10 @@ const mockStore = configureMockStore(middlewares)
 
 const rootOrganisationUnit = 'abc123'
 const relativePeriod = 'xyzpdq'
+const digitGroupSeparator = 'COMMA'
 selectors.sGetRootOrgUnit = () => rootOrganisationUnit
 selectors.sGetRelativePeriod = () => relativePeriod
+selectors.sGetSettingsDigitGroupSeparator = () => digitGroupSeparator
 
 jest.mock('../../modules/orgUnit', () => ({
     convertOuLevelsToUids: (ouLevels, vis) => vis,
@@ -177,6 +179,7 @@ describe('index', () => {
                     value: {
                         rootOrganisationUnit,
                         relativePeriod,
+                        digitGroupSeparator,
                     },
                 },
             ]
@@ -207,6 +210,7 @@ describe('index', () => {
                     value: {
                         rootOrganisationUnit,
                         relativePeriod,
+                        digitGroupSeparator,
                     },
                 },
             ]