+++ packages/app/src/actions/index.js
@@ -1,4 +1,5 @@
 import i18n from '@dhis2/d2-i18n'
+import { getDisplayNameByVisType } from '@dhis2/analytics'
 
 import {
     apiFetchVisualization,
@@ -19,7 +20,11 @@ import * as fromLoader from './loader'
 
 import { sGetCurrent } from '../reducers/current'
 import { sGetVisualization } from '../reducers/visualization'
-import { sGetRootOrgUnit, sGetRelativePeriod } from '../reducers/settings'
+import {
+    sGetRootOrgUnit,
+    sGetRelativePeriod,
+    sGetSettingsDigitGroupSeparator,
+} from '../reducers/settings'
 
 import history from '../modules/history'
 import { getVisualizationFromCurrent } from '../modules/visualization'
@@ -107,8 +112,15 @@ export const clearVisualization = (dispatch, getState, error = null) => {
 
     const rootOrganisationUnit = sGetRootOrgUnit(getState())
     const relativePeriod = sGetRelativePeriod(getState())
+    const digitGroupSeparator = sGetSettingsDigitGroupSeparator(getState())
 
-    dispatch(fromUi.acClear({ rootOrganisationUnit, relativePeriod }))
+    dispatch(
+        fromUi.acClear({
+            rootOrganisationUnit,
+            relativePeriod,
+            digitGroupSeparator,
+        })
+    )
 }
 
 export const tDoRenameVisualization = ({ name, description }) => (
@@ -182,9 +194,16 @@ export const tDoSaveVisualization = ({ name, description }, copy) => async (
             delete visualization.id
         }
 
-        if (name) {
-            visualization.name = name
-        }
+        visualization.name =
+            name ||
+            i18n.t('Untitled {{visualizationType}} visualization, {{date}}', {
+                visualizationType: getDisplayNameByVisType(visualization.type),
+                date: new Date().toLocaleDateString(undefined, {
+                    year: 'numeric',
+                    month: 'short',
+                    day: '2-digit',
+                }),
+            })
 
         if (description) {
             visualization.description = description